--- templates/channels/channel_claim_africas_talking.haml
@@ -1,2 +1,0 @@
--extends "channels/channel_claim_form.html"
-
