+++ templates/channels/types/africastalking/claim.haml
@@ -1,0 +1,16 @@
+-extends "channels/channel_claim_form.html"
+
+-block title
+  .medium-help.float-left{class:"{{ view.channel_type.icon }}"}
+
+  %h2.font_normalize.header-margin
+    {{ title }}
+
+
+-block extra-script
+  {{ block.super }}
+
+  :javascript
+    $(function(){
+      $("#id_country").select2();
+    });
